backend/requirements.txt
Flask
flask-cors
pymongo
SpeechRecognition
pydub
pyaudio
gtts
playsound
pygame
python-telegram-bot==13.15